requirements.txt
cloudflare
supabase
boto3
python-dotenv
opencv-python